feat: Abstract synthetic decorations to work with old decoration protocol
Abstract inlay hint provider to work with clients that don't support this feature
