+++ metals/src/main/scala/scala/meta/internal/metals/clients/language/ConfiguredLanguageClient.scala
@@ -155,6 +155,18 @@ final class ConfiguredLanguageClient(
     }
   }
 
+  override def refreshInlayHints(): CompletableFuture[Void] = {
+    if (clientConfig.isInlayHintsEnabled()) {
+      underlying
+        .refreshInlayHints()
+        .handle { (msg, ex) =>
+          if (ex != null)
+            scribe.warn(s"Error while refreshing inlayHints: $msg", ex)
+          msg
+        }
+    } else CompletableFuture.allOf()
+  }
+
   private def toShowMessageRequestParams(
       params: MetalsQuickPickParams
   ): ShowMessageRequestParams = {
