+++ mtags-shared/src/main/scala/scala/meta/internal/pc/LabelPart.scala
